Implement CLI formatter dual perc
